requirements.txt
accelerate>=0.19.0
braceexpand>=0.1.7
# data_utils>=0.0.7
einops>=0.6.1
einops_exts>=0.0.4
fastapi>=0.95.2
gradio>=3.33.1
huggingface_hub>=0.13.3
importlib_metadata>=6.6.0
inflection>=0.5.1
markdown2>=2.4.8
more_itertools>=9.1.0
nltk>=3.8.1
numpy>=1.23.5
open_clip_torch>=2.16.0
opencv_python_headless>=4.5.5.64
Pillow>=9.5.0
pycocoevalcap>=1.
pycocotools>=2.0.6
Requests>=2.31.0
scipy>=1.10.1
timm>=0.9.2
torch==2.0.1
torchvision>=0.15.1
tqdm>=4.65.0
transformers>=4.29.0
uvicorn>=0.22.0
webdataset>=0.2.48
xformers==0.0.20
deepspeed
prettytable
datasets
evaluate
wandb
scikit-learn
open-flamingo==2.0.1
